+++ app/models/concerns/grid_export.rb
@@ -25,12 +25,12 @@ module GridExport
         grid_group_variable.child_variables.includes(domain: :domain_options).each do |child_variable|
           if child_variable.variable_type == 'checkbox'
             child_variable.domain_options.each do |domain_option|
-              sorted_responses = grid_sort_responses_by_sheet_id_for_checkbox(grid_group_variable, child_variable, sheet_scope, sheet_ids, domain_option)
+              sorted_responses = grid_sort_responses_by_sheet_id_for_checkbox(grid_group_variable, child_variable, sheet_ids, domain_option)
               formatted_responses = format_responses(child_variable, raw_data, sorted_responses)
               csv << [grid_group_variable.name, child_variable.option_variable_name(domain_option)] + formatted_responses
             end
           else
-            sorted_responses = grid_sort_responses_by_sheet_id_generic(grid_group_variable, child_variable, sheet_scope, sheet_ids)
+            sorted_responses = grid_sort_responses_by_sheet_id_generic(grid_group_variable, child_variable, sheet_ids)
             formatted_responses = format_responses(child_variable, raw_data, sorted_responses)
             csv << [grid_group_variable.name, child_variable.name] + formatted_responses
           end
@@ -49,11 +49,11 @@ module GridExport
     end
   end
 
-  def grid_sort_responses_by_sheet_id_for_checkbox(grid_group_variable, variable, sheet_scope, sheet_ids, domain_option)
+  def grid_sort_responses_by_sheet_id_for_checkbox(grid_group_variable, variable, sheet_ids, domain_option)
     responses = \
       Response.joins(grid: :sheet_variable)
-              .where(sheet_variables: { sheet_id: sheet_scope.select(:id), variable_id: grid_group_variable.id })
-              .where(sheet_id: sheet_scope.select(:id), variable_id: variable.id)
+              .where(sheet_variables: { sheet_id: sheet_ids, variable_id: grid_group_variable.id })
+              .where(sheet_id: sheet_ids, variable_id: variable.id)
               .left_outer_joins(:domain_option)
               .where(domain_options: { id: domain_option.id })
               .order('sheet_id desc', 'grids.position').distinct
@@ -61,10 +61,10 @@ module GridExport
     grid_sort_responses_by_sheet_id(responses, sheet_ids)
   end
 
-  def grid_sort_responses_by_sheet_id_generic(grid_group_variable, variable, sheet_scope, sheet_ids)
+  def grid_sort_responses_by_sheet_id_generic(grid_group_variable, variable, sheet_ids)
     response_scope = \
       Grid.joins(:sheet_variable)
-          .where(sheet_variables: { sheet_id: sheet_scope.select(:id), variable_id: grid_group_variable.id })
+          .where(sheet_variables: { sheet_id: sheet_ids, variable_id: grid_group_variable.id })
           .where(variable_id: variable.id)
           .order('sheet_id desc', :position)
     responses = if variable.variable_type == 'file'
@@ -79,7 +79,7 @@ module GridExport
   end
 
   def grid_sort_responses_by_sheet_id(responses, sheet_ids)
-    sorted_responses = Array.new(sheet_ids.count)
+    sorted_responses = Array.new(sheet_ids.size)
     response_counter = 0
     current_sheet_position = nil
     last_sheet_id = nil